Move coop incubator, cleanup
